Fix solc version detection for --solc-solcs-(select|bin) case
Currently, the solc version is detected only once when starting, but
this may cause a mismatch between the detected version and the
compiler version that is then used. Move the version extraction
to the _run_solc(...) function to avoid this problem. Also add
support for passing an environment to get_version(...), this is
required when using compilers managed with solc-select.