Remove 1 contract per compilation unit restriction
This PR removes the assumption that a compilation unit will not have
multiple contracts with the same name. This PR contains a lot of
**breaking changes**

Multiple functions are removed in `CompilationUnit`:
- contracts_filenames
- contracts_absolute_filenames
- filename_of_contract
- absolute_filename_of_contract
- used_filename_of_contract

As far as I can tell these functions where not used.

Additionally the standard crytic-compile format was changed, and now
include a filename per contract. However we kept the backward
compatibility for artifacts previously generated