platform: *: fix command execution on Windows
Some of the command executions could fail on Windows with errors
such as the following:

    FileNotFoundError: [WinError 2] The system cannot find the file specified

This is because the program name used in the command (e.g. "npm")
is unqualified, and PATH lookup is not performed on Windows when
`shell=False` is used.

We can avoid this problem by converting the executable name to a
fully-qualified path by using `shutil.which` and passing it as
`executable` on the `Popen` call.