Improve Windows support (#248)
* utils: npm: fix path detection on Windows

Windows will throw ValueError and not OSError if the path is too long

    pathobj = WindowsPath('{"compilation_units": {"/home/monty/tob/tools/slither/tests/ast-parsing/function-0.6.0.sol": {"compiler":...unction abstractFunc2() public virtual override(C6) {}/n    function abstractFunc3() public virtual override {}/n}"}}')
    args = ()

        @functools.wraps(strfunc)
        def wrapped(pathobj, *args):
    >       return strfunc(str(pathobj), *args)
    E       ValueError: stat: path too long for Windows

    c:\hostedtoolcache\windows\python\3.6.8\x64\lib\pathlib.py:387: ValueError

* platform: archive: fix path checking on Windows

Windows may throw an exception if the path is too long, which is
the case when the string is not really a path. Catch it and try
decoding the JSON in that case.

* platform: *: fix command execution on Windows

Some of the command executions could fail on Windows with errors
such as the following:

    FileNotFoundError: [WinError 2] The system cannot find the file specified

This is because the program name used in the command (e.g. "npm")
is unqualified, and PATH lookup is not performed on Windows when
`shell=False` is used.

We can avoid this problem by converting the executable name to a
fully-qualified path by using `shutil.which` and passing it as
`executable` on the `Popen` call.

* utils: naming: use POSIX-style paths for relative paths

When looking up full paths, Windows-style paths may show up, but
standardize on using POSIX-style paths for everything else.

* platform: solc: remove extra prints

These were introduced in bb594de09 ("upgrade pylint and black for super-linter v4")
and currently break Slither's tests, which expect a certain output.

* workflows: ci: disable embark

It is currently broken, and tries to pull a repository via git://
which GH Actions now blocks.